Avoid unintended payment cancellation with the modal
Once the payment is started the payment popup opens and the app
shows a modal informing the user that the payment continues in the
popup.

The problem here is that closing the modal would abort the payment
and it's not very clear what aborts the payment. For example clicking
outside of the modal would, and also closing the modal via the "X" at the top.

Try to make this all more explicit by removing the "closing when clicking outside"
(by removing data-micromodal-close) and instead of showing a closing icon
show a button which clearly states that it's aborting the payment.