Don't show the restart warning when the modal is visible
In case we start a payment we open den modal and the popup and
the payment status switches to "started". This makes a warning appear
that the payment is already in process which is meant to be shown in case
the user refreshes the page.

In the normal case where the modal is visible because the payment is in process
we don't want to show this warning. So track the visibility and hide the warning.