Show the restart warning in case the payment failed or got cancelled
In case the payment gets cancelled by the user we might get into either
the failed or cancelled state. Currently we just show the normal
payment UI in that case and don't give the user any feedback.

Instead show the restart warning in that case, which includes text
that the payment might have failed, so the user knows that we know
that things went wrong and a restart might be the next step.