Extend the pending notification to give information in case things were canceled
In case the user closes the popup we get a pending state from the backend and will never
know if the user user still in the payment process or has canceled it.

Try to work around this by providing some helpful information for both cases and also
provide a link to return to the initiating page, so the payment process could be restarted
from the beginning if needed.