Port to new routing-url attribute
The app-shell (or similar) now provide a routing-url
attribute which we subscribe to. It provides the URL part
that comes after everything up to and including the activity URL,
so anything that is "owned" by the activity.

This ports all code parsing that to the new attribute, for example:

* for "select": extract the payment ID
* for "create": extract all the payment parameters
* for "return": extract the PSP data for forwarding to the API

Since the attribute is only available once we are connected to the DOM
things have to be delayed a bit until we read them.

Most changes are for the create flow:

* We don't really used the parameters for rendering, so just parse
  the URL when needed and don't register anything with lit.
* Before we start the payment request we need to parse it again to
  check if auth is needed, previously it was parsed in the constructor.

XXX: routing-url currently is "null" in case there is no routing url. Ideally
this should be fixed in the provider, but hack around that for now by adding
null checks everywhere.

XXX: There is still some hacky URL parsing going on for URLs passed to the API
for callbacks/return URLs. For this we need a different solution.