Port to new routing-base-url provided attribute
The app-shell now provides a routing-base-url which is the URL under
which the activity is reachable, which the activity can use
as a callback URL for the payment provider.

This removes the remaining URL hacks in mono.

Also properly escape things for URL building while at it, though
there shouldn't be a difference with the values we use.