Introduce a default view for when there is no view selected
In case the app is viewed as is without going through the payment
creation link we would try to create a payment which would fail
due to missing info and then show an error.

This is not really helpful, so in case there is no view selected
we do nothing and show a warning message, that the payment can't
be started from within the app.